# src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Song Manager API Doc",
    "description": "API documentation for Song Manager Service",
    "version": "1.0.0",
    "contact": {
      "name": "Alemayehu Mekonen",
      "email": "alemayehu.dev@gmail.com",
      "url": "https://alemayehu.vercel.app"
    }
  },
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/songs/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}